net/hinic3: Synchronize new NIC features and bug fixes
driver inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/IC5K0L?from=project-issue
CVE: NA

--------------------------------

Add optical module type adaptation.
Add ethtool fec function.
Add NIC XDP PASS function.
Add QoS persistence.
Add NIC out-of-band VLAN configuration function.
Fix NIC bifur bugs.
Fix performance test bugs.
Fix NIC RX bugs.
Fix NIC VF bugs.
Fix the bug of exiting when probing NIC.
Fix bond bugs.
Fix the bug of MAC address configuration failure in OS 6.6.
Fix the bug of reading NIC status incorrectly.
Fix and add some statistics.

+++ drivers/net/ethernet/huawei/hinic3/hinic3_profile.h
@@ -38,13 +38,12 @@ struct hinic3_prof_adapter {
 #define LLT_STATIC_DEF_SAVED
 #endif
 
-/*lint -save -e661 */
 static inline struct hinic3_prof_adapter *
 hinic3_prof_init(void *device, struct hinic3_prof_adapter *adap_objs, int num_adap,
 		 void **prof_attr)
 {
 	struct hinic3_prof_adapter *prof_obj = NULL;
-	u16 i;
+	int i;
 
 	for (i = 0; i < num_adap; i++) {
 		prof_obj = &adap_objs[i];
@@ -68,7 +67,6 @@ static inline void hinic3_prof_deinit(struct hinic3_prof_adapter *prof_obj, void
 		prof_obj->deinit(prof_attr);
 }
 
-/*lint -restore*/
 
 /* module-level interface */
 #ifdef CONFIG_MODULE_PROF
